refactor(test): migrate AI parsing and generation unit tests
- Rename `ai-roadmap.service.spec.ts` to `ai/ai.service.spec.ts`.
- Update `AiService` tests to mock and assert raw string responses.
- Migrate JSON parsing and validation tests to `roadmaps` and `onboarding` test suites.
- Fix unbound method linter warnings (`@typescript-eslint/unbound-method`) in test specs.
- Clean up unused test imports and mock configurations.
